Combined definition of all click commands in single file cli.py
